smart_locker_project/requirements.txt
# Core dependencies (required for both MacBook and Raspberry Pi)
Flask
Flask-SQLAlchemy
Werkzeug
bcrypt
pyserial
Flask-Babel

# Hardware dependencies (Linux/Raspberry Pi only)
# evdev
# pyusb

# API dependencies
PyJWT==2.8.0
Flask-CORS==4.0.0 